Break composed resource rendering out into its own file
It's shared by both the PT and PTF composers.
