+++ internal/controller/apiextensions/composite/composition_render.go
@@ -1,0 +1,134 @@
+/*
+Copyright 2022 The Crossplane Authors.
+
+Licensed under the Apache License, Version 2.0 (the "License"); you may not use
+this file except in compliance with the License. You may obtain a copy of the
+License at
+
+    http://www.apache.org/licenses/LICENSE-2.0
+
+Unless required by applicable law or agreed to in writing, software distributed
+under the License is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR
+CONDITIONS OF ANY KIND, either express or implied. See the License for the
+specific language governing permissions and limitations under the License.
+*/
+
+package composite
+
+import (
+	"context"
+
+	"k8s.io/apimachinery/pkg/util/json"
+	"sigs.k8s.io/controller-runtime/pkg/client"
+
+	"github.com/crossplane/crossplane-runtime/pkg/errors"
+	"github.com/crossplane/crossplane-runtime/pkg/meta"
+	"github.com/crossplane/crossplane-runtime/pkg/resource"
+
+	v1 "github.com/crossplane/crossplane/apis/apiextensions/v1"
+	env "github.com/crossplane/crossplane/internal/controller/apiextensions/composite/environment"
+	"github.com/crossplane/crossplane/internal/xcrd"
+)
+
+// An APIDryRunRenderer renders composed resources. It may perform a dry-run
+// create against an API server in order to name and validate the rendered
+// resource.
+type APIDryRunRenderer struct {
+	client client.Client
+}
+
+// NewAPIDryRunRenderer returns a Renderer of composed resources that may
+// perform a dry-run create against an API server in order to name and validate
+// it.
+func NewAPIDryRunRenderer(c client.Client) *APIDryRunRenderer {
+	return &APIDryRunRenderer{client: c}
+}
+
+// Render the supplied composed resource using the supplied composite resource
+// and template. The rendered resource may be submitted to an API server via a
+// dry run create in order to name and validate it.
+func (r *APIDryRunRenderer) Render(ctx context.Context, cp resource.Composite, cd resource.Composed, t v1.ComposedTemplate, env *env.Environment) error { //nolint:gocyclo // Only slightly over (11).
+	kind := cd.GetObjectKind().GroupVersionKind().Kind
+	name := cd.GetName()
+	namespace := cd.GetNamespace()
+
+	if err := json.Unmarshal(t.Base.Raw, cd); err != nil {
+		return errors.Wrap(err, errUnmarshal)
+	}
+
+	// We think this composed resource exists, but when we rendered its template
+	// its kind changed. This shouldn't happen. Either someone changed the kind
+	// in the template or we're trying to use the wrong template (e.g. because
+	// the order of an array of anonymous templates changed).
+	if kind != "" && cd.GetObjectKind().GroupVersionKind().Kind != kind {
+		return errors.New(errKindChanged)
+	}
+
+	if cp.GetLabels()[xcrd.LabelKeyNamePrefixForComposed] == "" {
+		return errors.New(errNamePrefix)
+	}
+
+	// Unmarshalling the template will overwrite any existing fields, so we must
+	// restore the existing name, if any. We also set generate name in case we
+	// haven't yet named this composed resource.
+	cd.SetGenerateName(cp.GetLabels()[xcrd.LabelKeyNamePrefixForComposed] + "-")
+	cd.SetName(name)
+	cd.SetNamespace(namespace)
+
+	for i := range t.Patches {
+		if err := Apply(t.Patches[i], cp, cd, patchTypesFromXR()...); err != nil {
+			return errors.Wrapf(err, errFmtPatch, i)
+		}
+		if env != nil {
+			if err := ApplyToObjects(t.Patches[i], env, cd, patchTypesFromToEnvironment()...); err != nil {
+				return errors.Wrapf(err, errFmtPatch, i)
+			}
+		}
+	}
+
+	// Composed labels and annotations should be rendered after patches are applied
+	meta.AddLabels(cd, map[string]string{
+		xcrd.LabelKeyNamePrefixForComposed: cp.GetLabels()[xcrd.LabelKeyNamePrefixForComposed],
+		xcrd.LabelKeyClaimName:             cp.GetLabels()[xcrd.LabelKeyClaimName],
+		xcrd.LabelKeyClaimNamespace:        cp.GetLabels()[xcrd.LabelKeyClaimNamespace],
+	})
+
+	if t.Name != nil {
+		SetCompositionResourceName(cd, *t.Name)
+	}
+
+	// We do this last to ensure that a Composition cannot influence controller references.
+	or := meta.AsController(meta.TypedReferenceTo(cp, cp.GetObjectKind().GroupVersionKind()))
+	if err := meta.AddControllerReference(cd, or); err != nil {
+		return errors.Wrap(err, errSetControllerRef)
+	}
+
+	// We don't want to dry-run create a resource that can't be named by the API
+	// server due to a missing generate name. We also don't want to create one
+	// that is already named, because doing so will result in an error. The API
+	// server seems to respond with a 500 ServerTimeout error for all dry-run
+	// failures, so we can't just perform a dry-run and ignore 409 Conflicts for
+	// resources that are already named.
+	if cd.GetName() != "" || cd.GetGenerateName() == "" {
+		return nil
+	}
+
+	// The API server returns an available name derived from generateName when
+	// we perform a dry-run create. This name is likely (but not guaranteed) to
+	// be available when we create the composed resource. If the API server
+	// generates a name that is unavailable it will return a 500 ServerTimeout
+	// error.
+	return errors.Wrap(r.client.Create(ctx, cd, client.DryRunAll), errName)
+}
+
+// RenderComposite renders the supplied composite resource using the supplied composed
+// resource and template.
+func RenderComposite(_ context.Context, cp resource.Composite, cd resource.Composed, t v1.ComposedTemplate, _ *env.Environment) error {
+	for i, p := range t.Patches {
+		if err := Apply(p, cp, cd, patchTypesToXR()...); err != nil {
+			return errors.Wrapf(err, errFmtPatch, i)
+		}
+	}
+
+	return nil
+}
